+++ ReleaseNotes.md
@@ -4,7 +4,13 @@
 
 
 ## TODO
-- Feature: Add version of `SqliteInMemory.CreateOptions` etc. that returns `DbContextOptionsBuilder<T>`
+
+
+## 1.9.0
+
+- Feature: Add EfCoreResetter to reset the Primary and Foreign keys for save/restore of data in a database.
+- Feature: Add EfCoreResetter to anonymise names, addesses etc.
+- Bug fix: fix issue #15, EfSchemaCompare: Error with default values on string properties.
 
 ## 1.8.0
 - Improvement: CompareEfSql now recoginises unique indexes provided by constraints (rather than via the normal SQL Index statement)